[SIEM] Refactor hosts navigation (#46118)
