+++ x-pack/legacy/plugins/siem/public/components/navigation/tab_navigation/index.test.tsx
@@ -11,7 +11,7 @@ import { TabNavigation } from './';
 import { TabNavigationProps } from './types';
 import { navTabs, SiemPageName } from '../../../pages/home/home_navigations';
 import { HostsTableType } from '../../../store/hosts/model';
-import { navTabsHostDetails } from '../../../pages/hosts/hosts_navigations';
+import { navTabsHostDetails } from '../../../pages/hosts/details/nav_tabs';
 import { CONSTANTS } from '../../url_state/constants';
 import { RouteSpyState } from '../../../utils/route/types';
