+++ x-pack/legacy/plugins/siem/public/pages/hosts/details/body.tsx
@@ -11,9 +11,9 @@ import { indicesExistOrDataTemporarilyUnavailable, WithSource } from '../../../c
 import { setAbsoluteRangeDatePicker as dispatchAbsoluteRangeDatePicker } from '../../../store/inputs/actions';
 import { scoreIntervalToDateTime } from '../../../components/ml/score/score_interval_to_datetime';
 import { Anomaly } from '../../../components/ml/types';
-import { getHostDetailsEventsKqlQueryExpression } from '../helpers';
+import { getHostDetailsEventsKqlQueryExpression } from './helpers';
 
-import { HostDetailsBodyComponentProps } from './type';
+import { HostDetailsBodyComponentProps } from './types';
 import { getFilterQuery, type, makeMapStateToProps } from './utils';
 
 const HostDetailsBodyComponent = React.memo<HostDetailsBodyComponentProps>(
