+++ x-pack/legacy/plugins/siem/public/pages/hosts/details/helpers.test.ts
