+++ x-pack/legacy/plugins/siem/public/pages/hosts/details/helpers.ts
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { escapeQueryValue } from '../../lib/keury';
+import { escapeQueryValue } from '../../../lib/keury';
 
 /** Returns the kqlQueryExpression for the `Events` widget on the `Host Details` page */
 export const getHostDetailsEventsKqlQueryExpression = ({
