+++ x-pack/legacy/plugins/siem/public/pages/hosts/hosts.tsx
@@ -30,11 +30,7 @@ import { SpyRoute } from '../../utils/route/spy_routes';
 import { FiltersGlobal } from '../../components/filters_global';
 
 import * as i18n from './translations';
-import {
-  navTabsHosts,
-  AnomaliesQueryTabBodyProps,
-  HostsComponentsQueryProps,
-} from './hosts_navigations';
+import { navTabsHosts } from './nav_tabs';
 import { hasMlUserPermissions } from '../../components/ml/permissions/has_ml_user_permissions';
 import { MlCapabilitiesContext } from '../../components/ml/permissions/ml_capabilities_provider';
 
@@ -51,9 +47,6 @@ interface HostsComponentDispatchProps {
     to: number;
   }>;
 }
-
-export type CommonChildren = (args: HostsComponentsQueryProps) => JSX.Element;
-export type AnonamaliesChildren = (args: AnomaliesQueryTabBodyProps) => JSX.Element;
 
 export type HostsQueryProps = GlobalTimeArgs;
