+++ x-pack/legacy/plugins/siem/public/pages/hosts/hosts_body.tsx
@@ -11,14 +11,15 @@ import { indicesExistOrDataTemporarilyUnavailable, WithSource } from '../../cont
 
 import { hostsModel, hostsSelectors, State } from '../../store';
 
-import { HostsComponentProps, CommonChildren, AnonamaliesChildren } from './hosts';
+import { HostsComponentProps } from './hosts';
+import { CommonChildren, AnomaliesChildren } from './navigation/types';
 import { scoreIntervalToDateTime } from '../../components/ml/score/score_interval_to_datetime';
 import { setAbsoluteRangeDatePicker as dispatchSetAbsoluteRangeDatePicker } from '../../store/inputs/actions';
 import { Anomaly } from '../../components/ml/types';
 
 interface HostsBodyComponentProps extends HostsComponentProps {
   kqlQueryExpression: string;
-  children: CommonChildren | AnonamaliesChildren;
+  children: CommonChildren | AnomaliesChildren;
 }
 
 const HostsBodyComponent = memo<HostsBodyComponentProps>(
