+++ x-pack/legacy/plugins/siem/public/pages/hosts/navigation/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export * from './anomalies_query_tab_body';
+export * from './authentications_query_tab_body';
+export * from './events_query_tab_body';
+export * from './hosts_query_tab_body';
+export * from './uncommon_process_query_tab_body';
